requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic-settings==2.0.3

python-jose==3.3.0
cryptography==41.0.4
passlib==1.7.4
bcrypt==4.0.1

SQLAlchemy==2.0.21
asyncpg==0.28.0
